src/test/javascript/spec/app/entities/form-component-values-mp-pi/form-component-values-mp-pi.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { take, map } from 'rxjs/operators';
import { FormComponentValuesMpPiService } from 'app/entities/form-component-values-mp-pi/form-component-values-mp-pi.service';
import { IFormComponentValuesMpPi, FormComponentValuesMpPi } from 'app/shared/model/form-component-values-mp-pi.model';

describe('Service Tests', () => {
  describe('FormComponentValuesMpPi Service', () => {
    let injector: TestBed;
    let service: FormComponentValuesMpPiService;
    let httpMock: HttpTestingController;
    let elemDefault: IFormComponentValuesMpPi;
    let expectedResult;
    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule]
      });
      expectedResult = {};
      injector = getTestBed();
      service = injector.get(FormComponentValuesMpPiService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new FormComponentValuesMpPi(0, 'AAAAAAA', 'AAAAAAA', 'AAAAAAA', 'AAAAAAA');
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);
        service
          .find(123)
          .pipe(take(1))
          .subscribe(resp => (expectedResult = resp));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject({ body: elemDefault });
      });

      it('should create a FormComponentValuesMpPi', () => {
        const returnedFromService = Object.assign(
          {
            id: 0
          },
          elemDefault
        );
        const expected = Object.assign({}, returnedFromService);
        service
          .create(new FormComponentValuesMpPi(null))
          .pipe(take(1))
          .subscribe(resp => (expectedResult = resp));
        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject({ body: expected });
      });

      it('should update a FormComponentValuesMpPi', () => {
        const returnedFromService = Object.assign(
          {
            componentType: 'BBBBBB',
            value: 'BBBBBB',
            name: 'BBBBBB',
            extValue: 'BBBBBB'
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);
        service
          .update(expected)
          .pipe(take(1))
          .subscribe(resp => (expectedResult = resp));
        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject({ body: expected });
      });

      it('should return a list of FormComponentValuesMpPi', () => {
        const returnedFromService = Object.assign(
          {
            componentType: 'BBBBBB',
            value: 'BBBBBB',
            name: 'BBBBBB',
            extValue: 'BBBBBB'
          },
          elemDefault
        );
        const expected = Object.assign({}, returnedFromService);
        service
          .query(expected)
          .pipe(
            take(1),
            map(resp => resp.body)
          )
          .subscribe(body => (expectedResult = body));
        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a FormComponentValuesMpPi', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
